+++ modules/audio/src/loader.js
@@ -9,13 +9,13 @@ function Loader(ctx, events) {
       return buffers[name];
     },
 
-    load: function(baseUrl, samples, config) {
-      const names = Object.keys(samples);
-      events.emit("audio.load-all", names);
+    load: function(baseUrl, clips, urls) {
+      const names = Object.keys(clips);
+      events.emit("/audio/load-all", names);
 
       const promises = names.map(function(name) {
-        const sample = samples[name];
-        const url = baseUrl + sample.filename + config.audioFileExt;
+        const clip = clips[name];
+        const url = urls[0].replace("{{filename}}", clip.audio.filename);
         return fetchLocalOrRemote(url)
           .then(function(response) {
             return response.arrayBuffer();
@@ -24,12 +24,12 @@ function Loader(ctx, events) {
             return ctx.decodeAudioData(audioData);
           })
           .then(function(buffer) {
-            events.emit("audio.loaded-file", name);
+            events.emit("/audio/file-loaded", name);
             buffers[name] = buffer;
           });
       });
       return Promise.all(promises).then(function(buffers) {
-        events.emit("audio.loaded-all");
+        events.emit("/audio/all-loaded");
         return buffers;
       });
     }
@@ -39,8 +39,8 @@ function Loader(ctx, events) {
 module.exports = Loader;
 
 function fetchLocalOrRemote(url) {
-  if (process.env.NODE_ENV !== "production") {
-    const local = url.replace(STORAGE, "http://localhost:3333/data/audiosets/");
+  if (false && process.env.NODE_ENV !== "production") {
+    const local = url.replace(STORAGE, "http://localhost:3333/data/audiosets");
     return fetch(local).catch(function() {
       return fetch(url);
     });