+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # CHANGELOG
 
+## 14/04/2017 [Dani 6h]
+
+* Lik03
+* Rename webapp to app-visuals and refactor app
+
+## 12/04/2017
+
+* [Dani] 6h
+  * Set migrations
+  * Lik03 audioset
+  * Browser app
+
 ## 7/4/2017
 
 * [Dani] 8h